+++ src/main/java/minecrafttransportsimulator/guis/instances/AGUIPanel.java
@@ -61,7 +61,7 @@ public abstract class AGUIPanel extends AGUIBase{
 		xOffset = setupEngineComponents(guiLeft, guiTop, xOffset);
 		
 		//Add instruments.  These go wherever they are specified in the JSON.
-		for(Byte instrumentNumber : vehicle.instruments.keySet()){
+		for(Integer instrumentNumber : vehicle.instruments.keySet()){
 			//Only add instruments that have an optionalPartNumber as those are on the panel.
 			if(vehicle.definition.motorized.instruments.get(instrumentNumber).optionalPartNumber != 0){
 				addInstrument(new GUIComponentInstrument(guiLeft, guiTop, instrumentNumber, vehicle));
